pyproject.toml
[project]
name = "personal-assistant-low-level"
version = "0.1.0"
description = "My personal assistant using the providers API / no-frameworks"
requires-python = ">=3.12"
dependencies = [
    "openai",
    "fastapi",
    "pydantic[email]",
    "uvicorn",
    "requests>=2.32.3",
    "python-dotenv",
    "todoist-api-python>=2.1.7",
    "google-auth>=2.37.0",
    "google-api-python-client>=2.158.0",
    "google-auth-oauthlib>=1.2.1",
    "google-auth-httplib2>=0.2.0",
    "sqlalchemy>=2.0.37",
    "alembic>=1.14.1",
    "asyncpg>=0.30.0",
    "pydantic-settings>=2.7.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]


[project.scripts]
main = "app.main:run"

[tool.pytest.ini_options]
asyncio_mode = "auto"
